+++ app/javascript/custom/editors.js
@@ -1,5 +1,5 @@
-jQuery(function() {
-  var select = jQuery("#editor_kind");
+$(function() {
+  var select = $("#editor_kind");
   select.on("change", function() {
     if (this.value == "board") {
       $("#editor_title").removeAttr("disabled");
