+++ app/javascript/custom/loading_jquery.js
@@ -1,0 +1,5 @@
+// JQuery
+import jQuery from "jquery";
+import "jquery-ujs";
+// make jQuery global
+window.$ = window.jQuery = jQuery;
