+++ app/javascript/custom/papers.js
