+++ config/importmap.rb
@@ -2,6 +2,8 @@
 
 pin "application", preload: true
 
+pin_all_from "app/javascript/custom", under: "custom"
+
 # Chartkick
 pin "chartkick", to: "chartkick.js"
 pin "Chart.bundle", to: "Chart.bundle.js"
